LayoutTests/platform/glib/imported/w3c/web-platform-tests/trusted-types/trusted-types-reporting-for-Document-execCommand-expected.txt

CONSOLE MESSAGE: Refused to connect to http://web-platform.test:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to http://web-platform.test:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to connect to http://web-platform.test:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.

PASS No violation reported for TrustedHTML.
PASS No violation reported (paste command).
PASS Violation report for plain string (insertHTML command).
